chore(migrations): clean unused and fix down

+++ packages/backend/database/migrations/2025_06_15_215307_alter_client_id_nullable_on_prestations_table.php
@@ -8,7 +8,7 @@ return new class extends Migration {
     public function up(): void {
         Schema::table('prestations', function (Blueprint $table) {
             $table->dropForeign(['client_id']);
-            $table->integer('client_id')->nullable()->change();
+            $table->unsignedBigInteger('client_id')->nullable()->change(); // TODO: vérifier le type lors du déploiement
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade');
         });
     }
@@ -16,7 +16,7 @@ return new class extends Migration {
     public function down(): void {
         Schema::table('prestations', function (Blueprint $table) {
             $table->dropForeign(['client_id']);
-            $table->integer('client_id')->nullable(false)->change();
+            $table->unsignedBigInteger('client_id')->nullable(false)->change(); // TODO: vérifier le type lors du déploiement
             $table->foreign('client_id')->references('id')->on('clients')->onDelete('cascade');
         });
     }
